Reorganize example server backend and frontend layout
