+++ examples/server/server/backend/__init__.py
@@ -1,0 +1,9 @@
+"""Backend modules for the demo WebAuthn server."""
+
+__all__ = [
+    "attachments",
+    "attestation",
+    "config",
+    "metadata",
+    "storage",
+]
