+++ examples/server/server/static/advanced-auth/auth-advanced.js
