+++ examples/server/server/static/advanced-auth/auth-debug.js
