+++ examples/server/server/static/advanced-auth/binary-utils.js
