+++ examples/server/server/static/advanced-auth/credential-display.js
