+++ examples/server/server/static/advanced-auth/credential-utils.js
