+++ examples/server/server/static/advanced-auth/display-utils.js
