+++ examples/server/server/static/advanced-auth/hints.js
