+++ examples/server/server/static/advanced-auth/json-editor.js
