+++ examples/server/server/static/advanced-auth/navigation.js
