+++ examples/server/server/static/advanced-auth/status.js
