+++ examples/server/server/static/advanced-auth/styles.css
