+++ examples/server/server/static/advanced-auth/webauthn-json.browser-ponyfill.js
