--- examples/server/server/static/mds.css
@@ -1,4 +1,0 @@
-@import url('mds/overview.css');
-@import url('mds/table.css');
-@import url('mds/detail.css');
-@import url('mds/modal.css');
