+++ examples/server/server/static/mds/mds-utils.js
