+++ examples/server/server/static/mds/mds.css
@@ -1,0 +1,4 @@
+@import url('styles/overview.css');
+@import url('styles/table.css');
+@import url('styles/detail.css');
+@import url('styles/modal.css');
