+++ examples/server/server/static/mds/styles/detail.css
