+++ examples/server/server/static/mds/styles/table.css
